feat(faq): Refactor FAQ data structure and enhance FAQ items
